refactor: restructure LLM service by separating registry and service logic
- Removed the existing LLM service implementation and created a new modular structure.
- Introduced `LLMRegistry` for managing model configurations and retrieval.
- Created `LLMService` to handle LLM calls with retries and circular fallback.
- Added an `__init__.py` to define the LLM package interface.
- Enhanced code organization for better maintainability and clarity.
